refactor(Direction): rename package
