+++ src/test/java/nextstep/ladder/domain/DirectionTest.java
@@ -1,6 +1,6 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.step.Direction;
+import nextstep.ladder.domain.direction.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
